* release/v7.2.1:
  Bump version
  Rename route parameter 'central_domain' to 'routeDomain'
  Rename central_domain to routeDomain and remove tenant_domain route parameters
  Rename route pattern central_domain to routeDomain and move to cortex/tenants module
  Fix session input old data issues
  Check before detaching accessareas if deleted entity was soft deleted

+++ src/Traits/Accessible.php
@@ -102,7 +102,8 @@ trait Accessible
         }
 
         static::deleted(function (self $model) {
-            $model->accessareas()->detach();
+            // Check if this is a soft delete or not by checking if `SoftDeletes::isForceDeleting` method exists
+            (method_exists($model, 'isForceDeleting') && ! $model->isForceDeleting()) || $model->accessareas()->detach();
         });
     }
